fix: prevent codename collisions across concurrent agents
Two root causes fixed:
1. hashStringToIndex used only 2 bytes (UInt16) of SHA-256 hash,
   giving 65K possible values — too few to avoid birthday collisions.
   Now uses 4 bytes (UInt32) for 4B+ possible values.

2. getStoredAgentIdentity fell back to shared global agent_identity
   for UUID-based keys, causing new processes to inherit the previous
   agent's identity. Now returns null for instance:* keys, forcing
   fresh codename generation.

+++ src/mcp/codenames.ts
@@ -144,7 +144,7 @@ export function normalizeAgentBaseName(input: string): string {
 
 export function hashStringToIndex(value: string, modulo: number): number {
   const digest = createHash("sha256").update(value).digest();
-  return digest.readUInt16BE(0) % modulo;
+  return digest.readUInt32BE(0) % modulo;
 }
 
 export function codenameFromIndex(index: number): { name: string; display_name: string } {
